список приходит

+++ app/src/main/java/ru/practicum/android/diploma/filters/domain/model/Industry.kt
@@ -1,6 +1,6 @@
 package ru.practicum.android.diploma.filters.domain.model
 
 class Industry(
-    val id: Double,
+    val id: Int,
     val name: String
 )
